apps/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai>=1.6.1,<2.0.0
langchain==0.0.350
langchain-openai==0.0.2
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9
bcrypt==4.1.1
PyJWT==2.8.0
pytest>=8.2.0,<10.0.0
pytest-asyncio>=0.21.0
httpx==0.25.2

